ex220 v2 inact
This router is ISP version router.
The internal firmware also treats it as EX221
I'm not sure which nearest public release version it matchs to.

I know I should have created new board for this but since I didn't plan
to upstream these changes due to lack of clarity on nearest version of
it.

Signed-off-by: Naresh Solanki <naresh.solanki.2011@gmail.com>